Fix yet another bug with external tree stealers.
The bug was occurring in the stealer splitting when the newly created
right stealer had 0 as the next element on the stack.
Whenever such a stealer was stolen subsequently directly stolen
without a chance to update its internal state, a subsequent split
could potentially lose one of the children by leaving a 0 below
the left stealer traversing the last branch.

Solved by always pushing a completed subtree to the right substealer
when splitting a stealer.